Update note_id data type and set content in edit_note.html
I decided that keys should be set as strings instead of a UUID
object. I felt that having to call "note_id.hex" was inelegant
and was an implementation detail that should not be exposed. I
had the following alternatives:

1. keep the id as stored as a UUID object, and add a
"serialize_id" call in the Storage layer. I think the best way
to implement this would be to create a typedef and trait and
require that the serialize_id is implemented for the NOTE_ID
type... but then I remembered this isn't rust.

2. change the id to a string, and store the UUID as the hex
based string. Simple, straightforward. Implementation
independent. I only have to worry about ID generation at
the start, so this makes sense

The goal of the id is a stable identifier for my notes. If I chose the
title as the note id, then I would cause problems if I ever wanted to
change the title.

The open question now is whether I keep the id as a hex string or
include dashes. I want to keep dashes so I can human read the
post id in the URL. This is because I am using the note-id as the
URL mechanism.

In addition, I updated the Jinja template to put the markdown note
content into the edit_note.html